Proof of concept: summarize after responding to user
The roleplaysinglemodel custom chat summary workflow now runs in this
order:

1. GetCurrentSummaryFromFile
2. Standard
3. FullChatSummary

After step 2 (during if streaming), the LLM's response is returned to the user, but the connection is kept open until step 3 completes.

I had to add an argument to `run_workflow`, because returning a
generator is only allowed on one code path. Maybe reusing `stream`
would suffice? SO comments say changing the return type based on an
argument is an anti-pattern and the function contract should be
stricter.

The changed order also means memories/summaries can use the
assistant's response. I added a hard coded length based guess to
determine if the last assistant message should be used, but this
should be determined by the workflow order.